Fix prod deployment of base and cloud-init volumes
Since `use_base_volume` is set to `true`, the `root` volume does not need a `pool` defined. If a `pool` is defined, it'll conflict and cause the following errors.

```bash
Error: Invalid index
  on .terraform/modules/ncn-h004-v001-kubernetes-worker/main.tf line 60, in resource "libvirt_volume" "base":
  60:   pool   = libvirt_pool.pool["default"].name
    |----------------
    | libvirt_pool.pool is object with 1 attribute "ephemeral"
The given key does not identify an element in this collection value.
Error: Invalid index
  on .terraform/modules/ncn-h004-v001-kubernetes-worker/main.tf line 75, in resource "libvirt_cloudinit_disk" "init":
  75:   pool = libvirt_pool.pool["default"].name
    |----------------
    | libvirt_pool.pool is object with 1 attribute "ephemeral"
The given key does not identify an element in this collection value.
```