Fix missing size
The missing size causes the following error:

```bash
Error: Operation failed

  on .terraform/modules/ncn-h004-v001-kubernetes-worker/main.tf line 70, in resource "libvirt_volume" "vol":
  70:   size           = try(each.value.size, var.default_volume_size) * pow(1024, 3)
    |----------------
    | each.value.size is null
    | var.default_volume_size is 21

Error during operation: argument must not be null.
```

Since the key exists but is empty, it is interpretted by Terraform as a String of "null" which tricks the `test` since it isn't empty.

The size was missing from PR #10.